feat: reorganize project structure with Poetry packaging
- Move modules to src/ directory structure
- Add Poetry configuration with simple commands
- Update README with dual installation options
- Create comprehensive documentation
- Fix imports and add structured returns

Breaking: Updated import paths and module structure
Features: poetry run folderly-ai and folderly-cli commands
