+++ data/folderly_activities.json
@@ -1,0 +1,117 @@
+[
+  {
+    "timestamp": "2025-07-30T11:54:52.704932",
+    "action": "file_deleted",
+    "details": {
+      "filename": "test_1.txt",
+      "path": "C:\\Users\\aayus\\Desktop\\test_1.txt"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T12:07:50.812810",
+    "action": "file_deleted",
+    "details": {
+      "filename": "document_1.pdf",
+      "path": "C:\\Users\\aayus\\Desktop\\document_1.pdf"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.212681",
+    "action": "file_modified",
+    "details": {
+      "filename": "FullStac.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\FullStac.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.225691",
+    "action": "file_modified",
+    "details": {
+      "filename": "JobTracker.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\JobTracker.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.281579",
+    "action": "file_modified",
+    "details": {
+      "filename": "let.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\let.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.318610",
+    "action": "file_modified",
+    "details": {
+      "filename": "MongoDBCompass.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\MongoDBCompass.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.358127",
+    "action": "file_modified",
+    "details": {
+      "filename": "Wondershare Filmora 13.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\Wondershare Filmora 13.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.394135",
+    "action": "file_modified",
+    "details": {
+      "filename": "Insomnia.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\Insomnia.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.421138",
+    "action": "file_modified",
+    "details": {
+      "filename": ".git.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\.git.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:02:50.502980",
+    "action": "file_modified",
+    "details": {
+      "filename": "Visual Studio Code.lnk",
+      "path": "C:\\Users\\aayus\\Desktop\\Visual Studio Code.lnk"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:03:27.428035",
+    "action": "file_deleted",
+    "details": {
+      "filename": "ddf",
+      "path": "C:\\Users\\aayus\\Desktop\\ddf"
+    }
+  },
+  {
+    "timestamp": "2025-07-30T14:10:57.624624",
+    "action": "folder_created",
+    "details": {
+      "filename": "New folder",
+      "path": "C:\\Users\\aayus\\Desktop\\New folder",
+      "type": "folder"
+    }
+  },
+  {
+    "timestamp": "2025-07-31T09:52:28.785009",
+    "action": "file_deleted",
+    "details": {
+      "filename": "New folder",
+      "path": "C:\\Users\\aayus\\Desktop\\New folder",
+      "type": "file"
+    }
+  },
+  {
+    "timestamp": "2025-07-31T09:52:29.362134",
+    "action": "file_deleted",
+    "details": {
+      "filename": "holahola",
+      "path": "C:\\Users\\aayus\\Desktop\\holahola",
+      "type": "file"
+    }
+  }
+]
