+++ pyproject.toml
@@ -1,0 +1,35 @@
+[tool.poetry]
+name = "folderly"
+version = "1.0.0"
+description = "Smart Desktop File Management with AI Integration"
+authors = ["Folderly Team <aayush@aayussh.com>"]
+readme = "README.md"
+packages = [{include = "src"}]
+
+[tool.poetry.dependencies]
+python = "^3.8.1"
+openai = "^1.0.0"
+watchdog = "^3.0.0"
+send2trash = "^1.8.0"
+python-dotenv = "^1.0.0"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.0.0"
+black = "^23.0.0"
+flake8 = "^6.0.0"
+
+[tool.poetry.scripts]
+folderly = "src.cli.cli:main"
+folderly-ai = "src.ai.ai_integration:chat_with_ai"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.black]
+line-length = 88
+target-version = ['py38']
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py"]
