+++ tests/__init__.py
@@ -1,0 +1,6 @@
+"""
+Tests for Folderly
+Contains all test files and test utilities.
+"""
+
+# This file makes the tests directory a Python package
